+++ .github/workflows/lint_and_test.yml
@@ -61,7 +61,7 @@ jobs:
           pip install -e .          
       - name: Test with pytest
         run: |
-          pytest py3gpp/*
+          pytest tests/*
 
   pypi:
     needs: build
