+++ py3gpp/nrCRCDecode.py
@@ -13,12 +13,3 @@ def nrCRCDecode(blkcrc, poly, mask=0):
     L = len(encoded) - len(blkcrc)
     crc = encoded[-L:]
     return encoded[:len(blkcrc) - L], _binToInt(crc)
-
-def test_nrCRCDecode():
-    testdata = [0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0]
-    data, crc_result = nrCRCDecode(testdata, '24C')
-    assert crc_result == 0
-
-    testdata = [1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0]
-    data, crc_result = nrCRCDecode(testdata, '24C')
-    assert crc_result != 0
