+++ py3gpp/nrCRCEncode.py
@@ -57,12 +57,3 @@ def nrCRCEncode(blk, poly, mask=0):
     for i in range(L):
         blksrc[len(blk) + i] = (crc >> (L - i - 1)) & 1
     return blksrc
-
-
-def test_nrCRCEncode():
-    test = np.zeros(32)
-    test[0] = 1
-    result = nrCRCEncode(test, "24C")
-    # fmt: off
-    assert np.all(result == [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0])
-    # fmt: on
