+++ tests/test_nrCRCDecode.py
@@ -1,0 +1,10 @@
+from py3gpp.nrCRCDecode import nrCRCDecode
+
+def test_nrCRCDecode():
+    testdata = [0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0]
+    data, crc_result = nrCRCDecode(testdata, '24C')
+    assert crc_result == 0
+
+    testdata = [1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0]
+    data, crc_result = nrCRCDecode(testdata, '24C')
+    assert crc_result != 0
