+++ tests/test_nrCRCEncode.py
@@ -1,0 +1,10 @@
+import numpy as np
+from py3gpp.nrCRCEncode import nrCRCEncode
+
+def test_nrCRCEncode():
+    test = np.zeros(32)
+    test[0] = 1
+    result = nrCRCEncode(test, "24C")
+    # fmt: off
+    assert np.all(result == [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0])
+    # fmt: on
